Fix apn overwrite issue
When a new apn is created with an existing apn name but
different apn type, existing apn is overwritten with new apn.
This is because of the unique constraint defined for carriers
table.
Fix is to add carrier name, type fields also to the unique
constraint, which would allow multiple apns to share same
apn with different types.

Change-Id: I802d8a5379edb914df8b537e18882e4d0e18a8c1
CRs-Fixed: 886638